streamlit-llm/requirements.txt
streamlit
openai
langchain
python-dotenv
streamlit-chat